refactor general.py scatter_flat and scatter_nd, and (for tf and torch) _parse_ellipsis
