+++ ivy/functional/backends/mxnet/general.py
@@ -93,3 +93,33 @@ def cumprod(x, axis=0, exclusive=False):
     for array_chunk in array_stack[1:]:
         new_array_list.append(new_array_list[-1] * array_chunk)
     return _mx.nd.concat(*new_array_list, dim=axis)
+
+
+
+# noinspection PyShadowingNames
+def scatter_flat(indices, updates, size=None, tensor=None, reduction='sum', dev=None):
+    if ivy.exists(tensor):
+        raise Exception('MXNet scatter_flat does not support scattering into an pre-existing tensor.')
+    if reduction == 'replace':
+        return _mx.nd.scatter_nd(updates, _mx.nd.expand_dims(indices, 0), [size]).copyto(_mxnet_init_context(default_device(dev)))
+    else:
+        raise Exception('MXNet scatter_flat currently only supports reduction mode "replace", but {} selected.'.
+                        format(reduction))
+
+
+# noinspection PyShadowingNames
+def scatter_nd(indices, updates, shape=None, tensor=None, reduction='sum', dev=None):
+    if ivy.exists(tensor):
+        raise Exception('MXNet scatter_flat does not support scattering into an pre-existing tensor.')
+    if dev is None:
+        dev = _callable_dev(indices)
+    shape = list(shape)
+    num_idx_dims = len(indices.shape)
+    transpose_order = [num_idx_dims-1] + list(range(num_idx_dims-1))
+    indices = _mx.nd.transpose(indices, transpose_order)
+    shape = shape if type(shape) is list else shape.asnumpy().astype(_np.int32).tolist()
+    if reduction == 'replace':
+        return _mx.nd.scatter_nd(updates, indices, shape).copyto(_mxnet_init_context(dev))
+    else:
+        raise Exception('MXNet scatter_nd currently only supports reduction mode "replace", but {} selected.'.
+                        format(reduction))
