+++ ivy/functional/ivy/old/general.py
@@ -466,55 +466,6 @@ def meshgrid(*xs: Iterable[Union[ivy.Array, ivy.NativeArray]], indexing: str = '
     return _cur_framework().meshgrid(*xs, indexing=indexing)
 
 
-# noinspection PyShadowingNames
-def scatter_flat(indices: Union[ivy.Array, ivy.NativeArray], updates: Union[ivy.Array, ivy.NativeArray],
-                 size: Optional[int] = None, tensor: Optional[Union[ivy.Array, ivy.NativeArray]] = None,
-                 reduction: str = 'sum', dev: ivy.Device = None)\
-        -> Union[ivy.Array, ivy.NativeArray]:
-    """
-    Scatter flat updates into a new flat array according to flat indices.
-
-    :param indices: Indices for the new values to occupy.
-    :type indices: array
-    :param updates: Values for the new array to hold.
-    :type updates: array
-    :param size: The size of the result.
-    :type size: int
-    :param tensor: The tensor in which to scatter the results, default is None, in which case the size is used to
-                    scatter into a zeros array.
-    :param reduction: The reduction method for the scatter, one of 'sum', 'min', 'max' or 'replace'
-    :type reduction: str
-    :param dev: device on which to create the array 'cuda:0', 'cuda:1', 'cpu' etc. Same as updates if None.
-    :type dev: ivy.Device, optional
-    :return: New array of given shape, with the values scattered at the indices.
-    """
-    return _cur_framework(indices).scatter_flat(indices, updates, size, tensor, reduction, dev)
-
-
-# noinspection PyShadowingNames
-def scatter_nd(indices: Union[ivy.Array, ivy.NativeArray], updates: Union[ivy.Array, ivy.NativeArray],
-               shape: Optional[Iterable[int]] = None, tensor: Optional[Union[ivy.Array, ivy.NativeArray]] = None,
-               reduction: str = 'sum', dev: ivy.Device = None)\
-        -> Union[ivy.Array, ivy.NativeArray]:
-    """
-    Scatter updates into a new array according to indices.
-
-    :param indices: Indices for the new values to occupy.
-    :type indices: array
-    :param updates: Values for the new array to hold.
-    :type updates: array
-    :param shape: The shape of the result. Default is None, in which case tensor argument must be provided.
-    :type shape: sequence of ints
-    :param tensor: The tensor in which to scatter the results, default is None, in which case the shape arg is used to
-                    scatter into a zeros array.
-    :param reduction: The reduction method for the scatter, one of 'sum', 'min', 'max' or 'replace'
-    :type reduction: str
-    :param dev: device on which to create the array 'cuda:0', 'cuda:1', 'cpu' etc. Same as updates if None.
-    :type dev: ivy.Device, optional
-    :return: New array of given shape, with the values scattered at the indices.
-    """
-    return _cur_framework(indices).scatter_nd(indices, updates, shape, tensor, reduction, dev)
-
 
 # noinspection PyShadowingNames
 def gather(params: Union[ivy.Array, ivy.NativeArray], indices: Union[ivy.Array, ivy.NativeArray], axis: int = -1,
